pyproject.toml
[project]
name = "ai-profit-hunter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "newspaper3k>=0.2.8",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "supabase>=2.28.3",
]
